+++ code/src/hooks/useDictionary.tsx
@@ -1,0 +1,12 @@
+"use client";
+import { useContext } from 'react';
+import { DictionaryContext } from '../utils/dictionary/DictionaryContext';
+
+
+export const useDictionary = () => {
+    const context = useContext(DictionaryContext);
+    if (!context) {
+        throw new Error('useDictionary must be used within a DictionaryProvider');
+    }
+    return context;
+};
